add a workaround to avoid a gcc internal compiler error when using ruby 1.9.2-p290 on ubuntu 11.10 (64bit). (reported by Bob Saveland.)

+++ ext/oci8/ocinumber.c
@@ -1329,7 +1329,7 @@ static VALUE onum_hash(VALUE self)
 {
     char *c  = DATA_PTR(self);
     int size = c[0] + 1;
-    int i, hash;
+    long i, hash;
 
     /* assert(size <= 22); ?*/
     if (size > 22)
